+++ semgrep_rules_check/react_rules/positives.tsx
@@ -1,0 +1,13 @@
+<div>
+  <div dangerouslySetInnerHTML={{ __html: "HTML" }}>
+    Children
+  </div>
+
+  <Hello dangerouslySetInnerHTML={{ __html: "HTML" }}>
+    Children
+  </Hello>
+
+  <div dangerouslySetInnerHTML={{ __html: "HTML" }} />
+
+  <Hello dangerouslySetInnerHTML={{ __html: "HTML" }} />
+</div>