+++ semgrep_rules_check/run_check.sh
@@ -1,4 +1,11 @@
 #!/usr/bin/env sh
 
 ./run_semgrep_json.sh > received.json
+
+negatives=$(grep -n negatives received.json)
+if [ "$negatives" != "" ]; then
+  printf "$negatives\n"
+  exit 1
+fi
+
 diff -u expected.json received.json