+++ semgrep_rules_check/run_semgrep_json.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env sh
 
-semgrep -f ../rules --json . --strict --no-rewrite-rule-ids | python -mjson.tool
+semgrep -f ../rules --json . --strict --no-rewrite-rule-ids | python sort_n_print_json.py